changed the home page to the real home and got the free trial link working in the header
