+++ app/styles/components/_header.scss
@@ -8,7 +8,7 @@
   z-index: 200;
 
   background: linear-gradient(rgba(0, 0, 0, 0.3), rgba(0, 0, 0, 0.3)),
-    url("../../assets/media/images/header_og.jpg");
+    url("https://static1.squarespace.com/static/56103c12e4b02dbd233e1376/t/5ad02a6870a6adbfd8865c12/1526253766126/ban.jpg?format=2500w");
 
   background-repeat: no-repeat;
   background-size: 100%;
